+++ packages/cli/src/submit/stdin/Stdin.definition.ts
@@ -29,6 +29,13 @@ export const SubmitStdinDefinition: ICommandDefinition = {
         },
     ],
     options: [
+        {
+            name: "encoding",
+            aliases: ["ec"],
+            description: "The remote encoding to use for the JCL content. Defaults to IBM-1047 if not specified.",
+            type: "string",
+            required: false,
+        },
         // {
         //     name: "wait",
         //     aliases: ["w"],